High-order type checking and simpler equality
The type-checker now operates on high-order terms directly. This greatly
improves its performance, and removes the need for shift/subst.

Equality now doesn't use UnionFind. Instead, it just returns true when
the hash of the two terms are equal, or when we re-visit an equation.

Terms uses Bruijn levels instead of Bruijn indices. This removes the
need for `bind_free_vars`, and simplifies other functions.

This speeds up type-checking of moonad/moonad by 5x, and should decrease
the real complexity of the language considerably. That's old approach
needed more functions (shift/subs) and dependend on words, hashes and an
UnionFind structure. A further improvement would be to remove low-order
terms entirely, making the parser return high-order terms.